# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for AllianceCup Application",
        "title": "AllianceCup API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/admin/all-order": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get order by status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api/admin"
                ],
                "summary": "Get Orders",
                "operationId": "get order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Last item created at for pagination",
                        "name": "created_at",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by order status",
                        "name": "order_status",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/order.Order"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
...
        "/api/admin/processed-order": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "handler for admin/moderator to processed order by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api/admin"
                ],
                "summary": "Processed order by id",
                "operationId": "processed order",
                "parameters": [
                    {
                        "description": "order status",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ProcessedOrderStatus"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
...
        "/api/client/get-order": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "gets user order full info by its id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api/client"
                ],
                "summary": "GetOrderById",
                "operationId": "gets order by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "order id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/order.FullInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
...
        "/api/client/user-order": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "gets user order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api/client"
                ],
                "summary": "GetUserOrders",
                "operationId": "gets order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "last order created_at for pagination",
                        "name": "created_at",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/order.FullInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
...
        "/api/new-order": {
            "post": {
                "description": "creates a new order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "NewOrder",
                "operationId": "creates an order",
                "parameters": [
                    {
                        "description": "order info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/order.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.OrderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
        "/api/order-info-types": {
            "get": {
                "description": "get payment and delivery types",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "Get order info types",
                "operationId": "get order info types",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/shopping.DeliveryPaymentTypes"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handler.OrderResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.ProcessedOrderStatus": {
            "type": "object",
            "required": [
                "orderId",
                "toStatus"
            ],
            "properties": {
                "orderId": {
                    "type": "string"
                },
                "toStatus": {
                    "type": "string"
                }
            }
        },
...
        "order.FullInfo": {
            "type": "object",
            "properties": {
                "delivery": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/order.OrdersDelivery"
                    }
                },
                "info": {
                    "$ref": "#/definitions/order.Order"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/order.ProductFullInfo"
                    }
                }
            }
        },
        "order.Order": {
            "type": "object",
            "required": [
                "delivery_type_title",
                "order_sum_price",
                "payment_type_title",
                "user_email",
                "user_firstname",
                "user_lastname",
                "user_middle_name",
                "user_phone_number"
            ],
            "properties": {
                "closed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "delivery_type_title": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "order_comment": {
                    "type": "string"
                },
                "order_status": {
                    "type": "string"
                },
                "order_sum_price": {
                    "type": "number"
                },
                "payment_type_title": {
                    "type": "string"
                },
                "user_email": {
                    "type": "string"
                },
                "user_firstname": {
                    "type": "string"
                },
                "user_lastname": {
                    "type": "string"
                },
                "user_middle_name": {
                    "type": "string"
                },
                "user_phone_number": {
                    "type": "string"
                }
            }
        },
        "order.OrdersDelivery": {
            "type": "object",
            "properties": {
                "delivery_description": {
                    "type": "string"
                },
                "delivery_title": {
                    "type": "string"
                }
            }
        },
        "order.ProductFullInfo": {
            "type": "object",
            "properties": {
                "amount_in_stock": {
                    "type": "number"
                },
                "article": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "img_url": {
                    "type": "string"
                },
                "order_id": {
                    "type": "string"
                },
                "packages_in_box": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "price_for_quantity": {
                    "type": "number"
                },
                "product_title": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "units_in_package": {
                    "type": "integer"
                }
            }
        },
...
        "shopping.DeliveryPaymentTypes": {
            "type": "object",
            "properties": {
                "deliveryTypes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/shopping.DeliveryType"
                    }
                },
                "paymentTypes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/shopping.PaymentType"
                    }
                }
            }
        },
        "shopping.DeliveryType": {
            "type": "object",
            "properties": {
                "delivery_type_title": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "shopping.PaymentType": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "payment_type_title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}